Move cache module from sync to utils
